# docs/swagger.yaml
basePath: /v1
definitions:
...
  entity.Profile:
    properties:
      age:
        type: integer
      first_name:
        type: string
      full_name:
        type: string
      last_name:
        type: string
      middle_name:
        type: string
      sex:
        type: string
      user_id:
        type: string
    required:
    - age
    - first_name
    - last_name
    - sex
    type: object
...
  v1.errorResponse:
    properties:
      message:
        type: string
      ok:
        type: boolean
    type: object
...
  v1.statusResponse:
    properties:
      ok:
        type: boolean
    type: object
info:
  contact: {}
  description: REST API example
  title: Go Orders API Test Issue
  version: "1.0"
paths:
...
  /profiles:
    post:
      consumes:
      - application/json
      description: Create profile for user
      operationId: profile-create
      parameters:
      - description: profile data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/entity.Profile'
      produces:
      - application/json
      responses:
        "200":
          description: id
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create profile
      tags:
      - profile
  /profiles/{id}:
    get:
      consumes:
      - application/json
      description: get profile
      operationId: profile-get
      parameters:
      - description: Profile ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Profile'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get profile
      tags:
      - profile
    put:
      consumes:
      - application/json
      description: update profile
      operationId: profile-update
      parameters:
      - description: Profile ID
        in: path
        name: id
        required: true
        type: string
      - description: profile data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/entity.Profile'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.statusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update profile
      tags:
      - profile
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"